+++ src/pages/products.vue
@@ -1,33 +1,15 @@
 <template>
-  <v-container class="pt-0">
-    <CategorySelector />
-    <v-row>
-      <v-col
-        v-for="product in products"
-        :key="product.id"
-        cols="4"
-      >
-        <ProductCard :info="product" />
-      </v-col>
-    </v-row>
-  </v-container>
+  <ErrorBoundary>
+    <Suspense>
+      <template #default>
+        <ProductList />
+      </template>
+      <template #fallback>
+        <LoadingSpinner />
+      </template>
+    </Suspense>
+  </ErrorBoundary>
 </template>
 
 <script setup>
-import api from '@/apis/config';
-
-const products = ref(null);
-const route = useRoute();
-const query = route.query;
-
-async function fetchDefaultData() {
-  try {
-    const res = await api.get(`/api/v1/products?typeId=${query.typeId}`);
-    products.value = res.data;
-  } catch (error) {
-    throw new Error(error);
-  }
-}
-
-onMounted(fetchDefaultData);
 </script>
